+++ shuup/admin/modules/categories/views/list.py
@@ -21,7 +21,8 @@ class CategoryListView(PicotableListView):
     default_columns = [
         Column("image", _("Image"), sortable=False, linked=True, raw=True),
         Column(
-            "name", _(u"Name"), sortable=False, display="format_name", linked=True, allow_highlight=False,
+            "name", _(u"Name"), sort_field="translations__name", display="format_name",
+            linked=True, allow_highlight=False,
             filter_config=MPTTFilter(
                 choices="get_name_filter_choices",
                 filter_field="id"