+++ CHANGELOG.md
@@ -8,9 +8,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 List all changes after the last release here (newer on top). Each change on a separate bullet point line.
 
+
 ### Fixed
 
 - Admin: Notification name when deleteing it
+
+### Changed
+
+- Front: Add supplier choice to best selling product context function
+- Admin: allow sorting categories by name
+- Admin: show product orderability errors as list
+
 
 ## [1.10.9] - 2020-03-24