+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 List all changes after the last release here (newer on top). Each change on a separate bullet point line.
 
 
+## [1.10.10] - 2020-03-24
+
 ### Fixed
 
 - Admin: Notification name when deleteing it